Отработка замечаний
