+++ MovieQuizTests/Mocks/ArithmeticOperationsMock.swift
@@ -1,0 +1,22 @@
+import XCTest
+
+final class ArithmeticOperationsMock {
+    func addition(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+            handler(num1 + num2)
+        }
+    }
+    
+    func subtraction(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+            handler(num1 - num2)
+        }
+    }
+    
+    func multiplication(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+            handler(num1 * num2)
+        }
+    }
+}
+
