+++ MovieQuizTests/Mocks/MovieQuizViewControllerMock.swift
@@ -1,0 +1,28 @@
+import XCTest
+@testable import MovieQuiz
+
+final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
+    func show(quiz step: QuizStepModel) {
+        
+    }
+    
+    func show(quiz result: QuizResultsModel) {
+        
+    }
+    
+    func highlightImageBorder(isCorrectAnswer: Bool) {
+        
+    }
+    
+    func showLoadingIndicator() {
+        
+    }
+    
+    func hideLoadingIndicator() {
+        
+    }
+    
+    func showNetworkError(message: String) {
+        
+    }
+}
