+++ MovieQuizTests/MovieQuizTests.swift
@@ -1,28 +1,9 @@
 import XCTest
 
-struct ArithmeticOperations {
-    func addition(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            handler(num1 + num2)
-        }
-    }
-    
-    func subtraction(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            handler(num1 - num2)
-        }
-    }
-    
-    func multiplication(num1: Int, num2: Int, handler: @escaping (Int) -> Void) {
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            handler(num1 * num2)
-        }
-    }
-}
-class MovieQuizTests: XCTestCase {
+final class MovieQuizTests: XCTestCase {
     func testAddition() throws {
         // Given
-        let arithmeticOperations = ArithmeticOperations()
+        let arithmeticOperations = ArithmeticOperationsMock()
         let num1 = 1
         let num2 = 2
