+++ MovieQuizUITests/MovieQuizUITests.swift
@@ -1,6 +1,6 @@
 import XCTest
 
-class MovieQuizUITests: XCTestCase {
+final class MovieQuizUITests: XCTestCase {
     var app: XCUIApplication!
     
     override func setUpWithError() throws {
@@ -19,36 +19,38 @@ class MovieQuizUITests: XCTestCase {
     
     
     func testYesButton() {
+        // Given
         sleep(3)
         let firstPoster = app.images["Poster"]
         let firstPosterData = firstPoster.screenshot().pngRepresentation
         
+        // When
         app.buttons["Yes"].tap()
         sleep(3)
         
         let secondPoster = app.images["Poster"]
         let secondPosterData = secondPoster.screenshot().pngRepresentation
-        
         let indexLabel = app.staticTexts["Index"]
         
+        // Then
         XCTAssertNotEqual(firstPosterData, secondPosterData)
         XCTAssertEqual(indexLabel.label, "2/10")
     }
     
     func testNoButton() {
+        // Given
         sleep(3)
-        
         let firstPoster = app.images["Poster"]
         let firstPosterData = firstPoster.screenshot().pngRepresentation
         
+        // When
         app.buttons["No"].tap()
         sleep(3)
-        
         let secondPoster = app.images["Poster"]
         let secondPosterData = secondPoster.screenshot().pngRepresentation
-        
         let indexLabel = app.staticTexts["Index"]
         
+        // Then
         XCTAssertNotEqual(firstPosterData, secondPosterData)
         XCTAssertEqual(indexLabel.label, "2/10")
     }
@@ -69,19 +71,20 @@ class MovieQuizUITests: XCTestCase {
     }
     
     func testAlertDismiss() {
+        // Given
         sleep(2)
         for _ in 1...10 {
             app.buttons["No"].tap()
             sleep(2)
         }
-        
         let alert = app.alerts["Game results"]
         alert.buttons.firstMatch.tap()
         
+        // When
         sleep(2)
-        
         let indexLabel = app.staticTexts["Index"]
         
+        // Then
         XCTAssertFalse(alert.exists)
         XCTAssertTrue(indexLabel.label == "1/10")
     }
